requirements.txt
crewai
crewai-tools
streamlit
unstructured
tools
requests
fastapi
uvicorn
pydantic
python-dotenv
yfinance
